add 404 Not Found for not exist id

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Http\Request;
+use \Illuminate\Http\Response;
 
 /*
 |--------------------------------------------------------------------------
@@ -23,22 +24,30 @@ Route::post('products', 'ApiProductController@addProduct');
 
 
 Route::get('products/{product_id}', function ($product_id) {
+    if (!DB::table('products')->where('id', $product_id)->exists()) {
+        return response()->json([], Response::HTTP_NOT_FOUND);
+    }
     $product = \App\Product::find($product_id,['title','description','price']);
     return response()->json($product);
 });
 
 Route::put('products/{product_id}', function (Request $request, $product_id) {
+    if (!DB::table('products')->where('id', $product_id)->exists()) {
+        return response()->json([], Response::HTTP_NOT_FOUND);
+    }
     $product = \App\Product::find($product_id);
     $product->update($request->all());
     return response()->json($product);
 });
 
 Route::delete('products/{product_id}', function (Request $request, $product_id) {
+    if (!DB::table('products')->where('id', $product_id)->exists()) {
+        return response()->json([], Response::HTTP_NOT_FOUND);
+    }
     $product = \App\Product::find($product_id);
     $product->delete();
     return response()->json($product);
 });
-
 
 
 
@@ -48,11 +57,17 @@ Route::post('shops', 'ApiShopController@addShop');
 
 
 Route::get('shops/{shop_id}', function ($shop_id) {
+    if (!DB::table('shops')->where('id', $shop_id)->exists()) {
+        return response()->json([], Response::HTTP_NOT_FOUND);
+    }
     $shop = \App\Shop::find($shop_id,['name']);
     return response()->json($shop);
 });
 
 Route::put('shops/{shop_id}', function (Request $request, $shop_id) {
+    if (!DB::table('shops')->where('id', $shop_id)->exists()) {
+        return response()->json([], Response::HTTP_NOT_FOUND);
+    }
     $shop = \App\Shop::find($shop_id);
     $shop->update($request->all());
     return response()->json($shop);
@@ -60,6 +75,9 @@ Route::put('shops/{shop_id}', function (Request $request, $shop_id) {
 
 
 Route::delete('shops/{shop_id}', function (Request $request, $shop_id) {
+    if (!DB::table('shops')->where('id', $shop_id)->exists()) {
+        return response()->json([], Response::HTTP_NOT_FOUND);
+    }
     $shop = \App\Shop::find($shop_id);
     $shop->delete();
     return response()->json($shop);
